Port-StreamBase (#3)
* Initial implementation of Stream base classes.

* A round of fixes for clang.

Added warning supression for padding.  Added explicit casts for the signed to unsigned conversions.  Touched up some of the docs.

* Another round of fixes.

More of the same as the last commit, plus created some missing implementations.

* Added a warning supression for weak vtables.

Clang doesn't like that we don't have a source file with even one definition for the pure abstract classes that are the stream base interfaces.  So it doesn't know where to put the vtables for those classes...but it'll manage.

* Fixed a missing cast.

* Fixed a memory layout issue is MSVC.

* Added a series of Binary/Test Readers/Writers.

Also touched up small parts of the IOStream API to make them slightly cleaner.

* More code cleanup for Clang

Changed SizeType to StreamSize where applicable.  Added a couple exceptions.  Added a throw test.

* Two more casts because Clang is picky.

* Use numeric_limits instead of tricks.

* Another round of casts into SizeType.

* More casts, this time in TextStreamReader.

* Suppress dumb clang warning.

* Emscripten is last to go.

* Hopeful fix for *raspian.  Not emscripten.

* Minor code cleanup.